requirements.txt
flask
flask-wtf
flask-sqlalchemy
requests
flask-login
flask-migrate
pyjenkins
